requirements.txt
torch
torchvision
lightning
wandb
numpy
Pillow
fastapi
uvicorn
python-multipart
kagglehub
python-dotenv